+++ src/mcmc/MetropolisVariable.cpp
@@ -109,7 +109,7 @@ QMap<double, double> MetropolisVariable::generateHisto(const QVector<double>& da
     int outputSize = 2 * (inputSize / 2 + 1);
     
     double sigma = dataStd(dataSrc);
-    double h = 1.06f * sigma * powf(dataSrc.size(), -1.f/5.f);
+    double h = 1.06f * sigma * pow(dataSrc.size(), -1.f/5.f);
     double a = vector_min_value(dataSrc) - 4.f * h;
     double b = vector_max_value(dataSrc) + 4.f * h;
     double delta = (b - a) / fftLen;
@@ -128,7 +128,7 @@ QMap<double, double> MetropolisVariable::generateHisto(const QVector<double>& da
         for(int i=0; i<outputSize/2; ++i)
         {
             double s = 2.f * M_PI * i / (b-a);
-            double factor = expf(-0.5f * s * s * h * h);
+            double factor = exp(-0.5f * s * s * h * h);
             
             output[2*i] *= factor;
             output[2*i + 1] *= factor;
